Check whether a move really updated the state.
From RollingState, propagate back to callers whether the state was
actually changed with an on- or off-chain update.  In ChannelManager,
use that information (for off-chain updates only) to notify WaitForChange
and process resolutions only if there was actually a change.

In the future, we will also make use of that for auto moves and for
game-specific updates.  That will avoid calling those in situations
where there is not yet any current state, and it also in general
avoids false positives and thus makes processing more efficient.