Let games send on-chain moves for channel updates.
This adds another callback for OpenChannel, where games can implement their
own logic for sending on-chain moves in response to certain channel states
(apart from disputes/resolutions).

For Xayaships (and likely other games as well), this will be used to send
moves that close a channel in agreement when the off-chain board state
is finished and a signed winner statement provided.