Automoves for Xayaships.
This implements the logic to perform automoves with Xayaships.
In particular, automoves take care of all the commit-reveal parts of
the game protocol, of answering shots and of signing the winner statement
when the game is finished.

With this change done, all the "game logic" parts for Xayaships and its
OpenChannel implementation are finished.