Add unit test for throwing an exception.
This adds a trivial test binary that is built and run with "make check",
and simply throws and catches a C++ exception.  Since there are some
issues with exceptions on Windows
(https://github.com/xaya/libxayagame/issues/73), this might help to
debug the correct build flags there and also will catch any future
regressions in that respect.

Exceptions are an integral part of C++ and are used at the core of
libxayagame, e.g. through libjson-rpc-cpp.