Daemon class for all channel components.
The new class Daemon in the game-channel library ties together the various
components of a channel daemon.  It holds a ChannelManager instance as
well as the various things that feed updates into and out of it, like
the ChainToChannelFeeder, MoveSender and OffChainBroadcast.

It also is capable of running a blocking main loop similar to the Game
class from xayagame.