Binary and main function for ships-channel.
This adds a main function for the new ships-channel binary, which is the
channel daemon for Xayaships.  It mainly puts together the already
existing pieces for that (and does not have much of its own logic).

The binary also does not yet have a JSON-RPC interface of itself, that
will be added in a second commit.