Extended state JSON for Xayaships channels.
This adds more data to the state JSON returned for Xayaships:  The
player's own configuration of ships (if set), string representations
for potentially revealed ship positions, the winner (if set) as well as
the guesses (i.e., the positions and known_ships proto fields).

Having this data will allow frontends to work with the JSON without the
need to ever parse any protocol buffers.