Allow triggering automoves explicitly.
This adds a function to the ChannelManager's public interface that will
trigger automoves explicitly (if there are any).  It can be used by games
to trigger automoves in response to new user-provided input that may
enable them to be available.

For instance, if a user sets their position in Xayaships, we may then be
able to send an automove for the corresponding position commitment.