Store channel ID and meta in ParsedBoardState.
Previously, ProtoBoardState would keep a reference to the channel ID and
metadata (from BoardRules::ParseState).  This is very useful to have in
a parsed state, so it can be put into context.

This data is not specific to a proto-based state, though, and thus we
should store it with general ParsedBoardStates directly.