Use correct reference when parsing states.
Since a ParsedBoardState contains a reference to the channel metadata,
we have to be careful about using the right reference for ParseState;
in particular, it has to be around as long as the ParsedBoardState
instance itself.

In RollingState reinit entries, we do keep a copy of the channel metadata
in addition to a ParsedBoardState.  However, we were previously not
actually referencing that particular instance when parsing the state,
but a temporary one.  This caused dangling references.  In this change,
we make sure to use the correct reference, so that the lifecycle is
correct and no bad references occur anymore.