Keep track of on-chain block and height.
Store the on-chain block hash and height with each update to the
ChannelManager.  This allows us to know what exact on-chain block
the current state corresponds to, and return that in the state JSON.

This will be useful to avoid flaky behaviour in integration tests (as
we can make sure to wait for updates to propagate), and it will also
be useful in general for frontends.