Test reorgs in Xayaships channels.
The new reorg.py integration test verifies what happens if reorgs happen
on-chain while a channel is being played.  In particular, we check that
it is fine if the channel creation gets reorged, what happens if another
player joins due to a reorg and how the channel daemon handles reorgs of
resolutions, dispute timeouts and winner statements.