Add custom field to channel metadata.
The new field "custom" can be used by games to hold their own data
needed to determine channel details.  For instance, they can store
what participants agreed upon when opening the channel, e.g. the amount
of money the game is for, or some special rules / "parameters".

It can also be used in case of forks in the game logic to determine what
rules (pre- / post-fork) should be applied for this particular channel
(so it is fixed during creation).  This will be useful to tackle
https://github.com/xaya/libxayagame/issues/49 for board state protos.