Implement HasAnyUnknownFields.
Add a utility function that checks recursively if a protocol buffer
message contains any unknown fields (directly or in a nested submessage).

This will be used to enforce that protos parsed from external input do
not have any, in order to avoid issues with bad data in the blockchcain
that prevents future proto upgrades.  For more details, see
https://github.com/xaya/libxayagame/issues/49.