Enforce proto version for StateProofs.
When processing StateProofs in the game-channel library (for
disputes / resolutions in the GSP and off-chain updates on an open
channel), enforce version-checks and checks for no unknown fields.

This also exposes the full versioning check (including unknown fields)
for SignedData proto's, which may be used by games directly (e.g. for
winner statements in Xayaships).