Check signed winner-statement protos.
Apply versioning checks and checks for unknown fields for signed winner
statement protos in Xayaships.  In particular, the "outer" SignedData
proto is passed through the versioning check of the game-channel framework
(so the format for signatures can be upgraded in a controlled way).

For the inner winner statement, we just forbid unknown fields for now.
In case we upgrade the proto format (which seems unlikely), we can always
add ships-specific versioning checks in place easily.