Forbid unknown fields in protoboard.
When parsing protos as board states and moves in protoboard, consider
them invalid if they contain any unknown fields.  This is part of solving
https://github.com/xaya/libxayagame/issues/49.

Since board state and move protos are game-specific, we do not apply any
other versioning checks.  It is up to games to do that as needed in
IsValid and ApplyMoveProto.