Track txid's of pending disputes / resolutions.
When sending a dispute or resolution move, keep track of the associated
transaction ID in ChannelManager.  Previously, we just kept a flag of
whether there is a pending transaction.  With the txid, we can do more
precise logic in the future, e.g. compare it to the mempool after
new blocks to see if the tx actually got confirmed before clearing
the pending flag.