Implement MoveSender::IsPending.
This adds a new method IsPending to MoveSender, which can be used to query
Xaya Core's mempool to see if a given move transaction is (still) pending.

With this, we will be able to implement better control over when to resend
disputes / resolutions for https://github.com/xaya/libxayagame/issues/65.