Check mempool before resetting pending moves.
When a new block comes in, check the mempool before clearing the pending
dispute / resolution txids.  Only clear them if they are really gone from
the mempool.  This avoids resending the moves if the pending name_update's
were not confirmed in a block due to miner policy.

There still remains an edge case that is not perfectly handled, though:
If a mined resolution is detached due to a reorg, we resend it even though
the previous tx may have been put back into the mempool.  This would be
hard to fix, as we would need to keep track of resolution transactions
even beyond clearing the active dispute.  Hence it seems not worth the
trouble for avoiding a superfluous transaction in edge cases.