Fix ChannelManager WaitForChange tests.
This tweaks the tests for ChannelManager's WaitForChange to be more
robust and test what they should correctly.  In particular, we make
sure to verify that the waiter threads are done when they should be
(rather than just let them timeout and still succeed if they were not
actually signalled).

Also includes a fix to the tests for off-chain updates, which previously
were timing out because there was no current on-chain state and thus
the off-chain move was actually ignored rather than applied and notifying
the waiter threads.