Add function for pending moves to ChannelManager.
The new ChannelManager::ProcessPending function will be used to handle
StateProofs extracted from pending moves seen in the mempool.  They are
handled essentially like off-chain moves, except that we assume their
reinit is the current on-chain one.  (That might not be true if the move
or some other pending move changes the reinit on-chain when confirmed,
but it is mostly the case and then when it is not, we will simply
fail processing the move if it is now invalid.)