Explicitly wait for processing of waitforchange.
Instead of a generic "SleepSome", explicitly wait in some unit tests
until the fake waitforchange method returns so that we know the change
has been processed.

This includes a refactor of how the waiting and notifying of threads is
done, so that the code is easier to follow and also can be reused for
testing of calls including waitforpendingchange in the future.