Process pending moves in channel daemon.
Listen to and process pending moves in channel daemons.  In particular,
we process updates as per ChannelGame::PendingMoves, which records
StateProof's seen in pending moves.

The channel daemon will try to process those StateProofs to see if they
yield a newer state, and try to continue the game based on them.  This
allows faster recovery back to real-time game play in case e.g. a dispute
is filed because some broadcast message was missed or the like.

We can also use that for more testing of pending-move tracking in general
through the integration tests.