Introduce HttpRpcServer for tests.
The new class template wraps a (mock) RPC server instance together
with an HttpServer, HttpClient and the RPC client.  This saves us from
having all those instances every time in tests.

For now, only the game_tests are updated to use it as a first step
and to verify all works.  More tests will be changed in follow-ups.