Add a validateposition RPC to ships-channel.
This RPC simply exposes the logic for parsing and validating a position
given as string (as per setposition) to the frontend.  This RPC is
completely independent of any state, and also does not require any
locks at all.  So it is purely a way to share the validation logic
with the frontend.