Print block processing times as LOG(INFO).
Instead of only logging block processing times as VLOG(1), print them
as LOG(INFO).  This information is not too spammy (as it will be printed
only once per block, which other stuff is as well), and can be very useful
especially also with VLOG turned off (so we get real performance data).