Make Random instances movable.
Allow moving of Random instances.  When an instance is moved, its state
(i.e. sequence of future bytes) is transferred onto the target.  The old
instance is then left unseeded.

This allows code like "Random rnd = <get new instance>" and also
re-assigning to a Random variable in this way.  That will be useful for
implementing branched off Random instances in the future.