Reseed Random already when last byte is used.
Instead of reseeding a Random instance only when new bytes are needed,
to the reseeding already when the last bytes of the current blob have
been used up.

This does not change the stream of random bytes (so it is a safe change
to make).  But it gives us a stricter guarantee on the internal state,
namely that nextIndex will always be less (and not less-or-equal) than
uint256::NUM_BYTES.  As such the behaviour is better well-defined and
simpler to understand.