Implement branched-off Random instances.
The new function Random::BranchOff will yield a separate Random instance
that is seeded based on the current Random's state and some key string
(so we can branch off multiple instances at a single point in time).  The
original state is not modified by this.

With this feature, we can split a single long sequence of random bytes into
a hierarchy of sequences, so that e.g. operations can be run in parallel
with their own deterministic Random instances.

This implements https://github.com/xaya/libxayagame/issues/88.