Implement compression functions in xayautil.
This adds a pair of functions CompressData and UncompressData to
xayautils.  These functions can be used to apply simple compression
(raw deflate) if needed by games, while they are specifically designed
to make sure the uncompress logic will be stable with respect to what
data it accepts and rejects in order to maintain consensus.