SQLiteDatabase to encapsulate the sqlite* handle.
The new class SQLiteDatabase wraps (and manages) an sqlite* handle,
together with the prepared-statement cache.  (This functionality is
moved out of SQLiteStorage.)

This will allow us to have multiple SQLite connections in the future,
in particular one write and multiple read-only snapshots.  For now, it
is just an internal refactoring, though.