Implement read-only snapshots in SQLiteStorage.
Allow SQLiteStorage instances to create read-only snapshots of the current
database state (this only works for WAL-mode files, i.e. not in-memory).

Snapshots allow performing a couple of SELECT statements, while not blocking
the database for future updates.  They can be used e.g. to execute expensive
RPC calls in an SQLite-based game while not blocking other RPCs or
processing of incoming blocks.