Unblocked callback for Game::GetCustomStateData.
Release the lock on Game before calling the extraction callback in
GetCustomStateData, so that potentially expensive callbacks can run
in parallel and do not block the entire GSP unnecessarily.

There is now a third variant of Game::GetCustomStateData, which is passed
the lock and can thus control the release finely as necessary.  This will
be used in the future for SQLiteGame, where we need to obtain a database
snapshot first before releasing the lock for the actual callback.