Add getnullstate as "standard" GSP RPC.
Add the method "getnullstate" to the standard GSP RPC (and implement
it in Game itself).  This method just returns the GSP state (e.g.
block height and syncing state) without any game data.  This is useful
to check on the GSP health and connection state.

The method is already present with custom implementations in e.g. the
Taurion GSP, and will be useful for others as well.  Thus even though it
is simple to implement manually, it makes sense to have it "standardised".