Add --xaya_rpc_wait to default main.
Support a new optional feature in default main, namely to wait for the
Xaya Core RPC interface to be available on startup (rather than expecting
it to be available and failing if not).

This is useful if we want to run GSPs automatically together with Xaya Core,
e.g. on system startup or in a Docker Compose configuration.