Wait before shutting down a Game.
When stopping a Game instance, wait some time (100ms in this case) to
give e.g. just notified waiter threads and other RPC calls a good chance
to return successfully before we stop the GSP completely.

This also includes some tweaks to Mover's waitforchange.py integration test,
which is no longer as flaky as before with this.