Add compression of JSON to xayautil.
This adds two new methods, CompressJson and UncompressJson, to xayautil.
They work similar to CompressData and UncompressData, but work on JSON
values directly (objects and arrays are supported).

The JSON is serialised, compressed and base64-encoded (although that
should be considered an implementation detail).  When uncompressing,
it is validated that the resulting data can be parsed back as JSON
with various, rather strict options.

This is something useful for games, where some data is selectively too
large for the Xaya move limit (e.g. tactics in SME or long waypoint
paths in Taurion).  Having the functionality directly in libxayagame
ensures that games do not need to worry about the slight intricacies
in ensuring valid JSON parsing.