Prevent non-integer JSON literals.
Json::Value's return true from isInt (and friends) not only if the value
is a true integer, but also if it is a floating-point number that can
be represented exactly in an integer.  Thus means that numbers parsed from
JSON literals like "42.0" or "1.5e1" will be seen as integers.

For stricter rules in GSPs, it may make sense to avoid those situations,
and only accept true integer literals.  This adds a new method to
xayautil/jsonutils.hpp, which can do exactly this check, and should
(if desired) be used in conjunction with is(U)Int(64).

We also employ the new method for enforcing integer literals in amounts
for the non-fungible GSP.