Implement state-to-JSON extraction.
The new class StateJsonExtractor implements the logic to retrieve
the full game state or selected parts of it from the database as
JSON values.  This will form the basis of the custom RPC interface
later; for now it is simply implemented and tested, and used for
the general "game-state to JSON" method of libxayagame.