Avoid zero'ing of zlib buffer.
Use direct char[] instead of std::string for the internal zlib buffer
in the compression util functions.  This avoids an unnecessary zeroing
step of the buffer on creation, which might be much larger (the maximum
output size) than the actual data at hand.

For instance, this zeroing step was responsible for about 10% of Taurion
GSP CPU usage during the 3rd competition (version 0.3).  With this change,
this part of the profile disappears entirely.