Abstraction class for SQLite statements.
Instead of exposing sqlite3_stmt directory for prepared statements
from SQLiteDatabase, abstract it behind a wrapper class.  For now, this
just provides a "step" function, which is a bit nicer to use than
sqlite3_step directly.  In the future, this will also provide some
basic bind/extract functions for integers and strings, and will be
used to make the statement cache recursive and thread-safe.

The actual GSP codes in other parts of the repository have not yet
been updated accordingly, but will be next.