Make SQLite statement cache thread-safe.
This makes the cache for prepared statements inside SQLiteDatabase
thread-safe, and also allows using the same statement multiple
times in parallel from a single thread.

Instead of just having a naive string->statement map, we use a multimap
that may have more than one cache entry for a particular SQL string.
Each entry also has an atomic flag, which indicates whether or not
it is currently in use.  This allows us to choose a free one, or create
another copy if none are free currently.

We make use of SQLiteDatabase::Statement as RAII wrapper to release
the locks on cached statements when they are no longer needed.