Implement "schema version" for SQLiteGame.
This adds a new concept/feature to SQLiteGame:  Game implementations
can read and write a "schema version" (which is just a string).  With this,
they can e.g. check what version previously wrote a database during
SetupSchema, and may use that to upgrade the schema / format as needed
when the GSP gets updated.