Allow testing broadcasts without ChannelManager.
This allows constructing OffChainBroadcast instances without a ChannelManager.
With this, it is easier to test the broadcasting mechanism itself of
a particular subclass, without the need to fake the manager.

We use this to simplify the RpcBroadcast test, but it will also be useful
in general for external broadcasting implementations (e.g. xmppbroadcast).