Move for declaring loss in Xayaships.
This implements a new move in Xayaships, which either channel participant
of an active game can use to just close the channel and declaring themselves
as loser.  This will replace the winner statement as the method for closing
a channel in agreement (it is simpler).

To declare loss, the move is simply:

  {"l": {"id": "CHANNEL-ID"}}