Loser declaration as auto move.
Instead of signing a winner statement and broadcasting it with the
auto-move framework in Xayaships, leave the game state at "winner
determined" and broadcast a loser declaration in this state.

In the future, we will also update the board rules to completely
eliminate the winner statement, in which case the final state of
the board will be no-turn again already when the winner is
determined; until then, there are some hacks required to the tests.