Remove winner statement completely.
This removes all traces of the winner statement from Xayaships, including
from the board rules, auto moves and on-chain move processing (for closing
a channel with a winner statement).

The board game concludes now (switches to a no-turn situation) as soon as
the winner has been determined.  At this step, it is expected (and
implemented in the default auto-move rules) that the loser declares loss
on-chain to close the channel.  If they don't, then the winner can use the
dispute/resolution mechanism to force-close.  (The latter does not work 100%
right now, but will be fixed soon.)