Add ChannelManager::PutStateOnChain.
Add a utility function to ChannelManager, which can be used to send a
resolution move explicitly (not necessarily in response to a dispute)
in order to put the currently best state on chain.  This can be useful
for some games in certain situations.

To implement this, we also keep track of the turn count of the best
state known on chain for each reinit we track.  With this, we can make
sure to only send the state on chain if it is actually better than
what is already known.