Force-close channels with filedispute.
This extends the "filedispute" RPC method of a Xayaships channel daemon
to put the current state on chain and force-close the channel if there
is already a winner determined.

This is needed to make sure the winner of a game can get this recorded
on chain even if the loser does not send the loss declaration as they
are supposed to do automatically.

Using the filedispute method for this purpose makes sense, since from
the frontend's point of view, this is very similar to a dispute (nothing
happens, and we need to enforce an action from the counterparty), even
though technically it is something different.