Pass through return value in AccessDatabase.
SQLiteDatabase allows external code to access the underlying sqlite3*
handle in a callback, so that locking is handled properly.  This small
change updates the template to allow the callback to return a value,
which will be passed back to the caller of AccessDatabase or
ReadDatabase accordingly.  That can be useful sometimes.