Callback-based default main for SQLite games.
This implements a new SQLiteMain overload, where the game logic is
defined by callbacks (that operate on raw sqlite3* handles) instead
of through an SQLiteGame subclass.

With this, it will be possible to implement a wrapper library
that allows implementing SQLite-based games in other languages
(e.g. C# similar to the existing wrapper for non-SQLite games).