Option for Xaya JSON-RPC protocol in channel.
This introduces a new command-line option for ships-channel (as well
as the underlying option in the general game-channel daemon API) which
can be used to switch the Xaya RPC to JSON-RPC 2.0 instead of the
legacy 1.0 that is needed for Xaya Core.  With this, it can work
properly with e.g. Electrum-CHI on light clients.