Generalise zmqPorts argument for xaya.Node.
Directly pass the -zmqpub* keys with ports in zmqPorts to a xaya.Node
instance (for the Python tests), instead of custom keys specific to
game blocks and pending moves.

This allows to configure the node also to publish raw blocks / hashes
instead of (or in addition to) the Xaya-specific endpoints.  With this,
we can test Xaya X more easily.