Generalise base chain for test framework.
This refactors the test framework (xayagametest), so that the base
chain that is linked to the GSP is generalised and run as an
"environment" by an overridable method.

For now, this does not change any functionality and still just runs a
Xaya Core instance.  But it makes it easier to replace the base chain
e.g. by a Xaya X instance in the future or in individual tests.