Support unspecified game genesis hash.
This extends the Game syncing logic so that games can optionally
leave the genesis hash unspecified.  In that case, the block hash
returned for the genesis height by the blockchain daemon will be
used instead.

With this, it is possible to run games e.g. on test blockchains that
do not have a fixed genesis hash, for instance Ganache.  For mainnets
it is still recommended to specify both height and hash, to ensure the
game starts at the right block.