Handle burn moves in xaya.py.
Move the explicit support for burn options in the test framework
from the generic testcase.py to xaya.py.  This is a move option that
is Xaya Core specific, and thus should be handled directly in the
Xaya Core environment.  The generic test fixture should just accept
and forward arbitrary arguments.