Support array pending moves.
This extends the handling of pending moves to support also JSON arrays
of individual JSON objects, i.e. transactions that trigger more than
one move (for instance, on an EVM-like base chain).

It also supports replacement of pending transactions:  If we receive
another notification for some txid and the associated data does not
match, we replace it and trigger a Reset of the pending state.

Fixes https://github.com/xaya/libxayagame/issues/112.