Build jsoncpp directly in source folder.
Instead of building outside the source tree, build jsoncpp in the
Dockerfile directly in the source folder.  This works as well and is
simpler, leading to a less confusing workflow (especially if the
Dockerfile is used as reference for how to build libxayagame on
a GNU/Linux system in general).