Bump library versions in Dockerfile.
Update the Dockerfile to use newer library versions of dependencies that
are installed from source:

  - jsoncpp: 1.9.4 -> 1.9.5
  - cppzmq: 4.7.1 -> 4.8.1
  - SQLite: 3.35.5 -> 3.36.0

A side effect of this is that we can get rid of the cherry-pick hack
that was required on jsoncpp 1.9.4.  Also due to newer versions, we need
to do an out-of-tree build again for jsoncpp (it fails otherwise).