Clarify processing order for ZMQ requests.
When the Game class calls game_sendupdates to request ZMQ updates,
it may happen that the backend actually starts pushing those messages
before the RPC call itself returns.

Due to the way in which Game handles mutex locks, this is fine already
in the code.  But this commit adds a comment explaining this situation
and why it works.