Abstract message signing in game channels.
This introduces two interfaces (SignatureVerifier and SignatureSigner),
which abstract message signing done in the game-channel code away from
directly using the Xaya Core RPCs.  This allows for making the underlying
game-channel code "more pure", and also e.g. to implement other signing
schemes and wallets, for instance for Ethereum-based chains.

With this change, we introduce implementations of those interfaces still
based on the Xaya Core RPC endpoints with verifymessage/signmessage, so
that functionality is not changed.  But now the direct use of the RPC
connection is pushed up from "basically all code" to just the top-level
class (ChannelGame for the GSP and Daemon for the channel-daemon), where
the RPC-based implementations are instantiated.

The design choice made for the signer is that it is aware of a particular
address that it can sign for (as it is tied to a particular participant).
This means that now the participant's signing address needs to be passed
when starting the channel daemon in addition to just the player name
(as the channel may not yet be on-chain initially so it can't be taken
from there), but this is a minor issue.