Generalise the MoveSender for game channels.
This splits out actual blockchain transaction sending from MoveSender
to a new class / interface TransactionSender.  The latter can implement
the basic operation of sending a move with given name and (string) value
as well as checking for pending transactions, while the former holds the
logic to construct the move data for disputes and resolutions.

The existing logic using Xaya Core's name_update RPC is provided
with the RpcTransactionSender.

In the gamechannel unit tests, we use direct mocks instead (and thus
remove the dependency on mocked RPC servers).  This will also allow
to reuse the game-channel logic in more general contexts in the future,
e.g. on Polygon.