Fix flakiness of channel waitforchange test.
The ships "waitforchange" channeltest was flaky due to race conditions
(where we sent "shoot" notifications to a channel daemon but then did
not necessarily wait for them to be processed before continuing with
the test).

This change introduces proper waiting routines, which explicitly
block until the turn count has increased as expected.