Remove Xaya RPC from channel Daemon.
This removes all logic specifically using a Xaya Core RPC connection
from the channel Daemon class.  Instead, the class uses explicitly
passed in SignatureVerifier, SignatureSigner and TransactionSender
instances (references).

The construction of those based on an RPC connection is now done directly
in the channel-daemon main file.  (But it could be something entirely
different, not based on JSON-RPC at all.)