Split receiving parts out of OffChainBroadcast.
The OffChainBroadcast class has two roles -- sending of messages and
receiving of them (and feeding them into ChannelManager::ProcessOffChain).
This change splits the receiving bits out, so that OffChainBroadcast
itself is just an abstract class for sending, i.e. the bits used directly
by the ChannelManager.

The event loop for receiving messages is moved to a subclass
(ReceivingOffChainBroadcast), so that the core bits needed for the
ChannelManager do not involve any threading.