Split gamechannel/gamestatejson.
The "gamestatejson" file contains code related both to pure channel
states (e.g. metadata or board states to JSON) as well as the database
representation of channels from the GSP.

This change splits the file up, so that the parts without database
dependency are moved to channelstatejson and can be used in a more general
context, while the database-code remains in gamestatejson for use in
the Xaya GSP backing channel games.