Split "channelcore" off game-channel library.
This splits the gamechannel library into two parts:  The new channelcore
library contains just the basics for managing a channel and state proofs,
and has minimal dependencies for that task.  In particular, it does not
depend on any networking, RPC or threading, and also not on SQLite or
libxayagame (just libxayautil).

This library can be used in restricted contexts, for instance to build
a channel-daemon in a web frontend using wasm.