Pass game ID to GetChannelSignatureMessage.
This modifies the game-channel code to pass through the game ID down
to GetChannelSignatureMessage.  For this, we add it to the ChannelManager
constructor at the top level (but that should not be an issue, as each
user will have a (likely fixed) game ID anyway).

For now, the value is then not actually used; but having it there will
allow us to include it in the signed messages in the future, for further
protection against relay attacks.