Add eth-utils as dependency.
Add the eth-utils library (from Xaya's Github repository) as dependency.
We will use it to provide an Ethereum-ECDSA-based signature signer and
verifier for game channels.

Note that this is useful even beyond applications built directly on an
Ethereum-like chain:  The signer is using libsecp256k1 directly, and
thus implements its own, self-contained signing/verification scheme,
which is quite useful in general (instead of relying on Xaya Core
RPCs for that).  It happens to be using the Ethereum signing methodology,
but can be applied in any channel application.