Queue off-chain moves for unknown reinits.
When we receive an off-chain update for an unknown reinit, queue it up
rather than discard it.  This way, we can apply it later on in case the
reinit becomes known.

This situation can happen if our peer receives an on-chain update faster
than us, reacts to it with e.g. some initial auto-move, and we receive it
before the corresponding on-chain update from our GSP.

In particular, this issue was also causing the reorg.py ships channeltest
to be flaky, which should be fixed now.  (But it is good to be fixed in
general, not just for the test, but something that can help in production
as well.)