DoS protection with off-chain update queue.
This adds some restrictions to the new queue of off-chain updates kept
for unknown reinitialisation states, to protect against DoS by peers
sending invalid updates.

In particular, the queue now has a limit on the total number of updates
kept (set to 100).  In practice, this limit should never be hit, as most
likely only one or two entries will be there at most (and even that only
in special edge cases).  But with this, the total memory footprint is
limited, so that a bad actor cannot use up all memory by sending lots of
large and invalid messages to the queue.