Generalise channeltest "runChannelDaemon".
This refactors the Python game-channel test fixture's handling of
how the channel daemon is run.  In particular, the general framework
assumes much less about the concrete arguments passed to the channel
daemon binary than before, while a custom override in the ShipsTest
class provides compatibility to the existing tests.

With this, it will be easier to support alternative things like running
on an EVM network in the future, where the arguments change compared
to running on Xaya Core.