Mention move id for UpdateMetadataReinit.
Clarify that the id passed to UpdateMetadataReinit should be a unique
identifier for the move, and that the transaction ID may not be such a
thing in some contexts (e.g. when running on an EVM chain, where a
transaction may trigger multiple moves, or trigger differing moves
depending on contract state).