Expose formatMove for integration tests.
This adds the utility method "formatMove" to the base class of
integration tests.  With this method, a move JSON can be formatted
into a string with the test's game ID.

Before, this was just exposed through the "sendMove" method, but it
can be useful to just format and then send manually.