Use mvid for ships channels if available.
Xaya X on EVM chains exposes a new field, mvid, which can be used
to better identify individual moves instead of the txid (which is not
necessarily unique to moves on EVM chains).

In Xayaships, we now use the mvid instead of txid for channel creation
(as the channel ID) and for updating a channel's reinit on joins,
if it is available.

This is a no-change on Xaya Core (where the txid itself is already a unique
identifier), but makes sure it will also work on EVM-chains with Xaya X.

Also contains some refactorings of the integration tests so that they
no longer assume channel id == txid, and instead take the channel ID from
the game state.