Use atomic for Game::state.
This changes the Game::state variable to be a std::atomic, and removes
the mutex lock when reading it for the health check (but all other
accesses, including all writes, still apply the mutex lock).

This makes the health check more responsive (e.g. while the GSP is
syncing or processing a new block).